api/requirements/local.txt
# Local development dependencies go here
django-mockups  #Genera datos de prueba
django-extensions #Agrega mas funcionalidad al shell de django
ipdb #debug el codigo ??? ipdb[TAB] en subl
#Requiere sudo aptitude install graphviz libgraphviz-dev pkg-configcl
pygraphviz #Genera imagen models ./manage.py graph_models -a -g -o my_project_visualized.png
